Fixed errors...
Changed double spin box to int, to match the slider. Connected those.
Changed the Connect Functions/Constructors so they are inside the main file bracket.

+++ mainwindow.h
@@ -29,7 +29,7 @@ public:
 
 private:
     Ui::MainWindow *ui;
-    unsigned double hw1, hw2, hw3, hw4, hw5, hw6, hw7, mt1, mt2, fnl;
+    double hw1, hw2, hw3, hw4, hw5, hw6, hw7, mt1, mt2, fnl;
 
 };
